[refactor] 04/05 닉네임 수정 컴포넌트 분리 #427
