+++ app/(main)/user/components/ProfileView.styled.tsx
@@ -59,31 +59,6 @@ export const ProfileContainer = styled.div`
   }
 `;
 
-export const NicknameContainer = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 0.25rem;
-`;
-
-export const NicknameBox = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 0.5rem;
-`;
-
-export const InputBox = styled.div`
-  width: 6rem;
-
-  input {
-    height: 100%;
-    padding: 0.3rem;
-    text-align: center;
-  }
-`;
-
 export const ErrorMessage = styled.p`
   font-size: var(--font-size-sm);
   color: var(--error-color);
